requirements.txt
tqdm==4.50.2
discord.py==1.5.0
buttons
argcomplete
